[probes] Move logic from __init__.py to __main__.py
__init__.py becomes the package marker only. ProbeResult, ProbeRunner, the
three probes, and main() all live in __main__.py — the file that `python -m
probes` runs. Script entry point and test imports updated to match.
